VP1-203-GrabableAndDoor
Fixed grabbable and door; door still needs a bit of cleanup, but overall scripting is complete. Separated default interaction into VLAT_GrabInteractable, similar to doors and drawers.
