+++ VLAT_Demo/Assets/VERA/ExampleInteractables/Scripts/Lock.cs.meta
