+++ VLAT_Demo/Assets/VERA/VLAT_LocomotionAccessibilityToolkit/VLAT_GrabInteractable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 58e85e2d31084e84a86c9386ee4c595e
+guid: 88e62d5c831599f4c98067d0df217077
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
